requirements.txt
gunicorn
boto3
Django==2.1.5
django-translations==1.1.0
django-polymorphic
django-storages
django-tinymce
martor
whoosh
django-haystack
unidecode
